Move mint to money actions.
